# docs/swagger.yaml
basePath: /api
definitions:
  api.JsonError:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  service.Review:
    properties:
      comment:
        example: 'Seriously... Who made this? '
        type: string
      created_at:
        example: "2025-10-30T00:00:00Z"
        type: string
      movie_id:
        example: 2
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  service.ReviewData:
    properties:
      comment:
        maxLength: 255
        type: string
      movie_id:
        type: integer
    required:
    - comment
    - movie_id
    type: object
...
  service.UpdatableReviewData:
    properties:
      comment:
        maxLength: 255
        type: string
    required:
    - comment
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Esta es la API REST para la aplicación de gestión de películas.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API de Películas UNICEN
  version: "1.0"
paths:
...
  /movies/{movieID}/reviews:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de reviews asociados a una película dada por ID,
        con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de reviews exitosa
          schema:
            items:
              $ref: '#/definitions/service.Review'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista los reviews de una película paginados
      tags:
      - Reviews
...
  /users/{userID}/reviews:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de reviews asociados a un usuario dada por ID,
        con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de reviews exitosa
          schema:
            items:
              $ref: '#/definitions/service.Review'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista los reviews de un usuario paginados
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Agrega una review de un usuario dado por ID a una película también
        dada por ID.
      parameters:
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      - description: Datos de la review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/service.ReviewData'
      produces:
      - application/json
      responses:
        "201":
          description: Review agregada exitosamente
          schema:
            $ref: '#/definitions/service.Review'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por review no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicación de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Agrega una review de un usuario sobre una película
      tags:
      - Reviews
  /users/{userID}/reviews/{movieID}:
    delete:
      consumes:
      - application/json
      description: Borra una review de un usuario dado por ID sobre una película dada
        también por ID permanentemente.
      parameters:
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Review borrada exitosamente
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por review no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Borra una review de un usuario sobre una película
      tags:
      - Reviews
    get:
      consumes:
      - application/json
      description: Obtiene una review de un usuario dado por ID sobre una película
        también dada por ID.
      parameters:
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review obtenida exitosamente
          schema:
            $ref: '#/definitions/service.Review'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por review no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Obtiene una review de un usuario sobre una película
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Actualiza una review de un usuario dado por ID sobre una película
        dada también por ID.
      parameters:
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      - description: Datos actualizables de la review
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/service.UpdatableReviewData'
      produces:
      - application/json
      responses:
        "200":
          description: Review actualizada exitosamente
          schema:
            $ref: '#/definitions/service.Review'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por review no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Actualiza una review de un usuario sobre una película
      tags:
      - Reviews
swagger: "2.0"